Revert "Make "Store" a trait"
This reverts commit 912c84e66347b082c82d91adc6b6da047a5160e4.
