+++ libimagstore/src/lib.rs
@@ -4,8 +4,6 @@ extern crate toml;
 pub mod content;
 pub mod entry;
 pub mod error;
-pub mod fsstore;
 pub mod header;
-pub mod single_use_lock;
 pub mod store;
