--- libimagstore/src/single_use_lock.rs
@@ -1,5 +1,0 @@
-pub trait SingleUseLock<T> {
-    fn access(&self) -> &T;
-    fn unlock(self) -> T;
-}
-
